feat: move variables to new package
