+++ tdp/cli/commands/restart_required.py
@@ -13,7 +13,7 @@ from tdp.cli.utils import check_services_cleanliness, collection_paths
 from tdp.core.dag import Dag
 from tdp.core.runner.ansible_executor import AnsibleExecutor
 from tdp.core.runner.operation_runner import EmptyOperationPlan, OperationRunner
-from tdp.core.service_manager import ServiceManager
+from tdp.core.variables import ServiceManager
 
 
 @click.command(short_help="Restart required TDP services")
