+++ tdp/cli/commands/run.py
@@ -11,7 +11,7 @@ from tdp.cli.utils import check_services_cleanliness, collection_paths
 from tdp.core.dag import Dag
 from tdp.core.runner.ansible_executor import AnsibleExecutor
 from tdp.core.runner.operation_runner import OperationRunner
-from tdp.core.service_manager import ServiceManager
+from tdp.core.variables import ServiceManager
 
 
 @click.command(short_help="Run single TDP operation")
