+++ tdp/core/models/service_log.py
@@ -6,7 +6,7 @@ from sqlalchemy.orm import relationship
 
 from tdp.core.models.base import Base
 from tdp.core.repository.repository import VERSION_MAX_LENGTH
-from tdp.core.service_manager import SERVICE_NAME_MAX_LENGTH
+from tdp.core.variables import SERVICE_NAME_MAX_LENGTH
 
 
 class ServiceLog(Base):
