+++ tdp/core/variables/__init__.py
@@ -1,0 +1,2 @@
+from .service_manager import SERVICE_NAME_MAX_LENGTH, ServiceManager
+from .variables import Variables, merge_hash
