+++ tdp/core/variables/cluster_variables.py
@@ -1,0 +1,5 @@
+from collections.abc import Mapping
+
+
+class ClusterVariables(Mapping):
+    pass
