+++ tdp/core/variables/service_manager.py
